Plugin manager refactored in sense of single responsibility principle.
